refactor(chat): move export options to AppChrome header menu
The per-chat export actions belong in the chat header's more-options
popover (AppChrome), not the project chat-list item menu. Move the two
LineItemButtons + handleExport there and revert ProjectChatSessionList.
The exportChatSession util and its tests are unchanged.
